+++ src/Slider.js
@@ -20,14 +20,15 @@ type Props = {
   handleClassName?: string,
   handleTitle?: React.Node,
   barClassName?: string,
-  hanldeStyle?: Object,
+  handleStyle?: Object,
   disabled?: boolean,
   graduated?: boolean,
   tooltip?: boolean,
   progress?: boolean,
   vertical?: boolean,
   onChange?: (value: number) => void,
-  renderMark?: (mark: number) => React.Node
+  renderMark?: (mark: number) => React.Node,
+  renderTooltip?: (value: number) => React.Node
 };
 
 type State = {
@@ -70,7 +71,7 @@ class Slider extends React.Component<Props, State> {
   getMouseMoveTracker() {
     return (
       this.mouseMoveTracker ||
-      new DOMMouseMoveTracker(this.hanldeDragMove, this.hanldeDragEnd, document.body)
+      new DOMMouseMoveTracker(this.handleDragMove, this.handleDragEnd, document.body)
     );
   }
 
@@ -151,7 +152,7 @@ class Slider extends React.Component<Props, State> {
     return value;
   }
 
-  hanldeClick = (event: SyntheticDragEvent<*>) => {
+  handleClick = (event: SyntheticDragEvent<*>) => {
     if (this.props.disabled) {
       return;
     }
@@ -166,7 +167,7 @@ class Slider extends React.Component<Props, State> {
   bar = null;
   handle = null;
 
-  hanldeMouseDown = (event: SyntheticEvent<*>) => {
+  handleMouseDown = (event: SyntheticEvent<*>) => {
     if (this.props.disabled) {
       return;
     }
@@ -181,14 +182,14 @@ class Slider extends React.Component<Props, State> {
     this.setTooltipPosition();
   };
 
-  hanldeDragEnd = () => {
+  handleDragEnd = () => {
     this.releaseMouseMoves();
     this.setState({
       handleDown: false
     });
   };
 
-  hanldeDragMove = (deltaX: number, deltaY: number, event: SyntheticDragEvent<*>) => {
+  handleDragMove = (deltaX: number, deltaY: number, event: SyntheticDragEvent<*>) => {
     if (!this.mouseMoveTracker || !this.mouseMoveTracker.isDragging()) {
       return;
     }
@@ -291,14 +292,22 @@ class Slider extends React.Component<Props, State> {
   }
 
   renderHanlde() {
-    const { handleClassName, handleTitle, min, vertical, tooltip, hanldeStyle } = this.props;
+    const {
+      handleClassName,
+      handleTitle,
+      min,
+      vertical,
+      tooltip,
+      handleStyle,
+      renderTooltip
+    } = this.props;
     const max = this.getMax();
     const { handleDown } = this.state;
     const value = this.getValue();
 
     const direction = vertical ? 'top' : 'left';
     const style = {
-      ...hanldeStyle,
+      ...handleStyle,
       // 通过 value 计算出手柄位置
       [direction]: `${(value - min) / (max - min) * 100}%`
     };
@@ -310,14 +319,14 @@ class Slider extends React.Component<Props, State> {
       <div
         className={handleClasses}
         role="presentation"
-        onMouseDown={this.hanldeMouseDown}
+        onMouseDown={this.handleMouseDown}
         onMouseEnter={this.handleMouseEnter}
         style={style}
         ref={this.handleRef}
       >
         {tooltip && (
           <Tooltip placement="top" className={this.addPrefix('tooltip')}>
-            {value}
+            {renderTooltip ? renderTooltip(value) : value}
           </Tooltip>
         )}
         {handleTitle}
@@ -362,7 +371,7 @@ class Slider extends React.Component<Props, State> {
     const unhandled = getUnhandledProps(Slider, rest);
 
     return (
-      <div {...unhandled} className={classes} onClick={this.hanldeClick} role="presentation">
+      <div {...unhandled} className={classes} onClick={this.handleClick} role="presentation">
         <div className={classNames(this.addPrefix('bar'), barClassName)} ref={this.barRef}>
           {progress && this.renderProgress()}
           {graduated && this.renderGraduated()}